Wtforms

+++ app.py
@@ -5,6 +5,8 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from models import Base, User, Entry, Tag
 from werkzeug.contrib.atom import AtomFeed
+from forms import LoginForm, CommentSubmit, BaseForm, Remove
+from wtforms import RadioField
 
 
 
@@ -45,8 +47,17 @@ def show_entries():
     if request.args.get('tag',''):
         entries = entries.filter(Entry.tag==request.args.get('tag',''))
     entries = entries.all()
+    tags = get_tags(db)
+    tags = [(str(tag[0]),str(tag[0])) for tag in tags]
+    form = CommentSubmit(meta={'csrf_context': session})
+    form.tags.choices = tags
+    remove = Remove(meta={'csrf_context': session})
+    return render_template('show_entries.html', entries=entries, tags=tags, form=form, remove=remove)
+
+def get_tags(db):
     tags = db.query(Tag.name).filter(Tag.username==session.get('username')).order_by(Tag.id.desc()).all()
-    return render_template('show_entries.html', entries=entries, tags=tags)
+    tags.append((u'Other',))
+    return tags
 
 @app.route('/signup', methods=['GET', 'POST'])
 def sign_up():
@@ -69,9 +80,13 @@ def sign_up():
 
 @app.route('/add', methods=['POST'])
 def add_entry():
+    db = get_db()
+    rform = CommentSubmit(request.form, meta={'csrf_context': session})
+    rform.tags.choices = [(str(tag[0]),str(tag[0])) for tag in get_tags(db)]
     if not session.get('logged_in'):
         abort(401)
-    db = get_db()
+    if not rform.validate():
+        abort(400)
     tag = ""
     if request.form['new_tag']:
         new_tag = Tag(name=request.form['new_tag'], username=session.get('username'))
@@ -88,8 +103,11 @@ def add_entry():
 
 @app.route('/remove', methods=['POST'])
 def remove_entry():
+    form = Remove(request.form, meta={'csrf_context': session})
     if not session.get('logged_in'):
         abort(401)
+    if not form.validate():
+        abort(400)
     db = get_db()
     db.delete(db.query(Entry).filter(Entry.id==request.args.get("id", '')).first())
     db.commit()
@@ -101,6 +119,9 @@ def login():
         return redirect(url_for('show_entries'))
     error = None
     if request.method == 'POST':
+        rform = LoginForm(request.form, meta={'csrf_context': session})
+        if not rform.validate():
+            abort(400)
         db = get_db()
         if not request.form['username']:
             error = 'Username is required'
@@ -119,7 +140,8 @@ def login():
     public_users = db.query(User.name).filter(User.public==1).all()
     public_users = [user[0] for user in public_users]
     entries = db.query(Entry.title, Entry.text, Entry.timestamp, Entry.id).filter(Entry.name.in_(public_users)).order_by(Entry.id.desc()).all()
-    return render_template('login.html', error=error, entries=entries)
+    form = LoginForm(meta={'csrf_context': session})
+    return render_template('login.html', error=error, entries=entries, form=form, meta={'csrf_context': request.session})
 
 @app.route('/logout')
 def logout():
